+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Moved default item builder to `DropDownItem.build`.
 * Added `DropDownItemBuilder` interface to allow the `T` of `DropDownItem<T>` to determine the widget displayed in the drop down list.
 * Moved search logic to `DropDownItem.satisfiesSearch` which is now called from `DropDownList.search`.
+* Added `DropDownItemSearchable` interface to allow the `T` of `DropDownItem<T>` to determine the way items are searched.
 
 ## 0.0.4